Move common UI components into UI (#5287)
This moves various UI components I've seen into the CLI's UI folder. It
moves select prompts and other things from utils into the UI if they are
related to a UI component.
