+++ ironfish-cli/src/commands/wallet/import.ts
@@ -11,7 +11,7 @@ import { Args, Flags, ux } from '@oclif/core'
 import { IronfishCommand } from '../../command'
 import { RemoteFlags } from '../../flags'
 import { inputPrompt } from '../../ui'
-import { importFile, importPipe, longPrompt } from '../../utils/input'
+import { importFile, importPipe, longPrompt } from '../../ui/longPrompt'
 import { Ledger } from '../../utils/ledger'
 
 export class ImportCommand extends IronfishCommand {
