+++ ironfish-cli/src/commands/wallet/multisig/commitment/create.ts
@@ -5,8 +5,7 @@ import { UnsignedTransaction } from '@ironfish/sdk'
 import { Flags } from '@oclif/core'
 import { IronfishCommand } from '../../../../command'
 import { RemoteFlags } from '../../../../flags'
-import { confirmOrQuit } from '../../../../ui'
-import { longPrompt } from '../../../../utils/input'
+import * as ui from '../../../../ui'
 import { MultisigTransactionJson } from '../../../../utils/multisig'
 import { renderUnsignedTransactionDetails } from '../../../../utils/transaction'
 
@@ -48,7 +47,7 @@ export class CreateSigningCommitmentCommand extends IronfishCommand {
 
     let identities = options.identity
     if (!identities || identities.length < 2) {
-      const input = await longPrompt(
+      const input = await ui.longPrompt(
         'Enter the identities of all participants who will sign the transaction, separated by commas',
         {
           required: true,
@@ -64,7 +63,7 @@ export class CreateSigningCommitmentCommand extends IronfishCommand {
 
     let unsignedTransactionInput = options.unsignedTransaction
     if (!unsignedTransactionInput) {
-      unsignedTransactionInput = await longPrompt('Enter the unsigned transaction', {
+      unsignedTransactionInput = await ui.longPrompt('Enter the unsigned transaction', {
         required: true,
       })
     }
@@ -81,7 +80,7 @@ export class CreateSigningCommitmentCommand extends IronfishCommand {
       this.logger,
     )
 
-    await confirmOrQuit('Confirm signing commitment creation', flags.confirm)
+    await ui.confirmOrQuit('Confirm signing commitment creation', flags.confirm)
 
     const response = await client.wallet.multisig.createSigningCommitment({
       account: flags.account,
