+++ ironfish-cli/src/commands/wallet/multisig/dkg/round3.ts
@@ -4,9 +4,7 @@
 import { Flags } from '@oclif/core'
 import { IronfishCommand } from '../../../../command'
 import { RemoteFlags } from '../../../../flags'
-import { inputPrompt } from '../../../../ui'
-import { longPrompt } from '../../../../utils/input'
-import { selectSecret } from '../../../../utils/multisig'
+import * as ui from '../../../../ui'
 
 export class DkgRound3Command extends IronfishCommand {
   static description = 'Perform round3 of the DKG protocol for multisig account creation'
@@ -47,12 +45,12 @@ export class DkgRound3Command extends IronfishCommand {
 
     let participantName = flags.participantName
     if (!participantName) {
-      participantName = await selectSecret(client)
+      participantName = await ui.multisigSecretPrompt(client)
     }
 
     let round2SecretPackage = flags.round2SecretPackage
     if (!round2SecretPackage) {
-      round2SecretPackage = await inputPrompt(
+      round2SecretPackage = await ui.inputPrompt(
         `Enter the round 2 encrypted secret package for participant ${participantName}`,
         true,
       )
@@ -60,7 +58,7 @@ export class DkgRound3Command extends IronfishCommand {
 
     let round1PublicPackages = flags.round1PublicPackages
     if (!round1PublicPackages || round1PublicPackages.length < 2) {
-      const input = await longPrompt(
+      const input = await ui.longPrompt(
         'Enter round 1 public packages, separated by commas, one for each participant',
         {
           required: true,
@@ -78,7 +76,7 @@ export class DkgRound3Command extends IronfishCommand {
 
     let round2PublicPackages = flags.round2PublicPackages
     if (!round2PublicPackages) {
-      const input = await longPrompt(
+      const input = await ui.longPrompt(
         'Enter round 2 public packages, separated by commas, one for each participant',
         {
           required: true,
