+++ ironfish-cli/src/ui/index.ts
@@ -4,6 +4,8 @@
 
 export * from './card'
 export * from './json'
+export * from './longPrompt'
 export * from './progressBar'
 export * from './prompt'
+export * from './prompts'
 export * from './table'
